refactor : actions 위치 이동
