+++ src/actions/gatherings/getMeetingDetail.ts
