+++ src/components/pages/findMeeting/FilterSort/FilterSort.tsx
@@ -2,7 +2,6 @@
 
 import { useRef, useState } from "react"
 
-import Sort from "@/components/public/icon/dynamicIcon/Sort"
 import { sortType } from "@/constants/meeting"
 import useOutsideClick from "@/util/useOutsideClick"
 
@@ -65,7 +64,6 @@ const FilterSort = (props: IFilterProps) => {
         onClick={onLabelClickHandler}
         onKeyDown={onLabelKeyDownHandler}
       >
-        <Sort state="default" />
         <span className="flex items-center whitespace-nowrap max-sm:hidden">
           {getSortType(selVal)}
         </span>
